refactor(OnTrackInstruction): Extract class from TripInstruction.
